third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-sizes-1x-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("simple").clientWidth is (window.innerWidth)
PASS document.getElementById("simple").currentSrc is "../../hidpi/resources/image-set-2x.png"
PASS document.getElementById("small_sizes").clientWidth is 300
PASS document.getElementById("small_sizes").currentSrc is "../../hidpi/resources/image-set-1x.png"
PASS document.getElementById("sizes_x_descriptors").clientWidth is 400
PASS document.getElementById("sizes_x_descriptors").currentSrc is "../../hidpi/resources/image-set-1x.png"
PASS document.getElementById("sizes_src").clientWidth is 400
PASS document.getElementById("sizes_src").currentSrc is "../../hidpi/resources/image-set-1x.png"
PASS document.getElementById("default_sizes").clientWidth is window.innerWidth
PASS document.getElementById("default_sizes").currentSrc is "../../hidpi/resources/image-set-2x.png"
PASS document.getElementById("invalid_length").clientWidth is window.innerWidth
PASS document.getElementById("invalid_length").currentSrc is "../../hidpi/resources/image-set-2x.png"
PASS document.getElementById("src_and_w").clientWidth is 700
PASS document.getElementById("src_and_w").currentSrc is "resources/image-set-4x.png"
PASS document.getElementById("src_and_w2").clientWidth is 700
PASS document.getElementById("src_and_w2").currentSrc is "../../hidpi/resources/image-set-1x.png"
PASS document.getElementById("x_and_w").clientWidth is 700
PASS document.getElementById("x_and_w").currentSrc is "resources/image-set-4x.png"
PASS document.getElementById("x_and_w2").clientWidth is 400
PASS document.getElementById("x_and_w2").currentSrc is "../../hidpi/resources/image-set-2x.png"
PASS document.getElementById("x_and_w3").clientWidth is 400
PASS document.getElementById("x_and_w3").currentSrc is "../../hidpi/resources/image-set-2x.png"
PASS document.getElementById("half_size").clientWidth is (window.innerWidth/2)
PASS document.getElementById("half_size").currentSrc is "../../hidpi/resources/image-set-2x.png"
PASS document.getElementById("infitisimal_sizes").clientWidth is (window.innerWidth/100)
PASS document.getElementById("infitisimal_sizes").currentSrc is "../../hidpi/resources/image-set-2x.png"
              
